+++ lib/pages/events_pages/events_page.dart
@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import '../../config/sort_filter_globals.dart';
-import 'event_listview.dart';
+import 'event_list_view.dart';
 
 final sortFilterStateKey = GlobalKey<SortFilterButtonState>();
