+++ lib/pages/events_pages/friends_footer.dart
@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:url_launcher/url_launcher.dart';
-import 'following_tweets_listview.dart';
+import 'following_tweets_list_view.dart';
 
 class FriendsFooter extends HookWidget {
   final Event _event;
